+++ __tests__/Clickable-tests.js
@@ -1,21 +1,18 @@
-import * as React from 'react';
-import { shallow, mount } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-import sinon from 'sinon';
+import React from 'react';
+import { mount } from 'enzyme';
 import Clickable from '../src/components/Clickable';
-import StopPropagation from '../src/components/StopPropagation';
 
 describe('Clickable component', () => {
   test('is clickable.', () => {
-    const onClick = sinon.spy();
-    const clickable = shallow(<Clickable onClick={onClick} />);
+    const onClick = jest.fn();
+    const clickable = mount(<Clickable onClick={onClick} />);
 
     clickable.simulate('click');
-    expect(onClick).toHaveProperty('callCount', 1);
+    expect(onClick).toHaveBeenCalledTimes(1);
   });
 
   test('onClick is called when child is clicked.', () => {
-    const onClick = sinon.spy();
+    const onClick = jest.fn();
     const html = mount(
       <Clickable onClick={onClick}>
         <div id="child" />
@@ -23,31 +20,6 @@ describe('Clickable component', () => {
     );
 
     html.find('#child').simulate('click');
-    expect(onClick).toHaveProperty('callCount', 1);
-  });
-});
-
-describe('StopPropagation component', () => {
-  test('stops event propagation to Clickable parent component.', () => {
-    const onClick = sinon.spy();
-    const html = mount(
-      <Clickable onClick={onClick}>
-        <StopPropagation>
-          <div id="child">
-            <div id="nestedChild" />
-          </div>
-        </StopPropagation>
-        <div id="clickableChild" />
-      </Clickable>
-    );
-
-    html.find('#child').simulate('click');
-    expect(onClick).toHaveProperty('callCount', 0);
-
-    html.find('#nestedChild').simulate('click');
-    expect(onClick).toHaveProperty('callCount', 0);
-
-    html.find('#clickableChild').simulate('click');
-    expect(onClick).toHaveProperty('callCount', 1);
+    expect(onClick).toHaveBeenCalledTimes(1);
   });
 });
