+++ package.json
@@ -22,7 +22,6 @@
     "enzyme": "^3.2.0",
     "enzyme-adapter-react-16": "^1.1.0",
     "jest": "^21.2.1",
-    "jest-enzyme": "^4.0.1",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "react-test-renderer": "^16.2.0",
@@ -31,8 +30,7 @@
     "rollup-plugin-babel": "^3.0.2",
     "rollup-plugin-commonjs": "^8.2.6",
     "rollup-plugin-flow": "^1.1.1",
-    "rollup-plugin-node-resolve": "^3.0.0",
-    "sinon": "^4.1.2"
+    "rollup-plugin-node-resolve": "^3.0.0"
   },
   "dependencies": {},
   "jest": {
